requirements.txt
streamlit>=1.50.0
pandas>=2.2.3
numpy>=1.26.4
requests>=2.32.3
python-dotenv>=1.0.1
plotly>=5.24.1
streamlit-aggrid>=1.1.8.post1
streamlit-plotly-events>=0.0.6
